Remove plugins - all broken since 34.11
This changes nothing at all as far as any DFHack user is concerned, as
these plugins have not been possible to build since DF 34.11 - and would
have to be rewritten for compatiblity with new native-DF systems.
"treefarm" is additionally replaced by "autochop".

In the unlikely event that a developer wants the source code for
something... that's what version control is for!