title-folder: fix SDL path on OS X
The Stonesense install script appears to do strange things which cause DF to
load SDL.framework/Versions/A/SDL instead of SDL.framework/SDL. Once the former
is loaded, loading the latter and calling SDL_WM_* functions from it will fail.

A better solution would be to remove parts of the Stonesense fix-libs-*.sh
script(s) that are causing this, since they don't appear to be needed, or
use dlsym() to find the already-opened library/symbols.