Switch to DFHack luacov fork
Fixes behavior of `push.recurseSubmodules`: since the commit we track is
currently only a tag (v0.15.0), pushing with `recurseSubmodules = check` fails
since that checks that submodule commits are on a branch. In our own fork, we
can add a `dfhack` branch that resolves this issue, which is consistent with
other dependencies.