update-submodules: hopefully fix pulling
This was the last failure. Unsure why this doesn't occur with other submodules.

https://github.com/DFHack/dfhack/runs/3307299302?check_suite_focus=true

```
+ git checkout dfhack
Previous HEAD position was 19b52ca 0.15.0 release
Switched to branch 'dfhack'
+ git pull --ff-only
There is no tracking information for the current branch.
Please specify which branch you want to merge with.
See git-pull(1) for details.

    git pull <remote> <branch>

If you wish to set tracking information for this branch you can do so with:

    git branch --set-upstream-to=origin/<branch> dfhack
```